Move more of log scanning into pure code
